login page backend, added sessions, added page redirects
In this commit:
(1) I add much of the backend code for the login page. Login page can now correctly verify a login attempt. However, in its current state, it cannot serve a user the set of authentication factors yet. But i can redirect an admin to the factors page.

(2) I added session variables. A session is started at the login page for all users and any variables placed in the session can be accessed later (global variables essentially)

(3) I added code for preventing page refresh. This code is added in the javascript. e.preventdefault

(4) I fixed the styling for buttons on the front end. The 3 file method for backend code doesnt work with <button> tag, so you have to use an <input> tag and disguise it as a bootstrap button by adding a class on it.

(5) I added the code/ability to redirect to different pages as needed. The idea should work for any page redirect.