Added backend functionality to Email Factor
Added the php code needed to retrieve the email address of the user/ companyID pair.

I noticed that there's a delay in retrieving/ displaying the email address from the database, but that may just be due to my current internet connection. Worse case we could store this in a session variable from the start and call it when needed.